+++ doc/development/releasenotes.rst
@@ -9,11 +9,14 @@ Development version in trunk
 ============================
 
 :trac:`trunk <>`.
-	  
+          
 * Genetic algorithm implemented; :mod:`ase.ga`. This can be used
   for the optimization of: atomic cluster structure, materials
   properties by use of template structures. Extension to other projects
   related to atomic simulations should be straightforward.
+
+* The :func:`ase.lattice.bulk` function can now build the Wurtzite structure.
+
 
 Version 3.8.0
 =============
@@ -128,7 +131,7 @@ Version 3.5.0
   * Povray render facility.
   * Updated expert user mode.
   * Enabled customization of colours and atomic radii.
-  * Enabled user default settings via :file:`~/.ase/gui.py`. 
+  * Enabled user default settings via :file:`~/.ase/gui.py`.
 
 * :mod:`Database library <data>` expanded to include: